Updated types and moved some helper functions to better locations
