+++ client/dashboard/emails/mappers/mailbox-to-email-mapper.ts
@@ -1,6 +1,6 @@
-import { Email, EmailAccount, EmailBox, SiteDomain } from '@automattic/api-core';
+import { EmailAccount, EmailBox, SiteDomain } from '@automattic/api-core';
 import { __ } from '@wordpress/i18n';
-import { accountHasWarningWithSlug } from '../utils/email-utils';
+import { accountHasWarningWithSlug } from '../../utils/email-utils';
 
 export const mapMailboxToEmail = (
 	box: EmailBox,
