+++ client/dashboard/emails/types.ts
@@ -1,0 +1,24 @@
+import { EmailProvider } from '@automattic/api-core';
+
+export interface Email {
+	id: string;
+	emailAddress: string;
+	type: 'mailbox' | 'forwarding';
+	provider: EmailProvider;
+	providerDisplayName: string;
+	domainName: string;
+	subscriptionId?: string;
+	siteId?: string;
+	siteName?: string;
+	forwardingTo?: string;
+	storageUsed?: number;
+	storageLimit?: number;
+	canUserManage: boolean;
+	status:
+		| 'active'
+		| 'pending'
+		| 'suspended'
+		| 'google_pending_tos_acceptance'
+		| 'unverified_forwards'
+		| 'unused_mailboxes';
+}
