+++ packages/api-core/src/emails/types.ts
@@ -7,29 +7,6 @@ export interface Mailbox {
 	mailbox: string;
 }
 
-export interface Email {
-	id: string;
-	emailAddress: string;
-	type: 'mailbox' | 'forwarding';
-	provider: EmailProvider;
-	providerDisplayName: string;
-	domainName: string;
-	subscriptionId?: string;
-	siteId?: string;
-	siteName?: string;
-	forwardingTo?: string;
-	storageUsed?: number;
-	storageLimit?: number;
-	canUserManage: boolean;
-	status:
-		| 'active'
-		| 'pending'
-		| 'suspended'
-		| 'google_pending_tos_acceptance'
-		| 'unverified_forwards'
-		| 'unused_mailboxes';
-}
-
 export interface DomainBinding {
 	domain: string;
 	is_primary: boolean;
